Make sure server dir doesn't already exist
Add helpful reminder to edit server.properties
Remove unneeded warning from README, it outlived its purpose by a few commits

+++ MCBEautoUpdate.sh
@@ -66,10 +66,15 @@ if [ -n "$service" ]; then
 		sudo systemctl start "$service"
 	fi
 else
+	if [ -d "$server_dir" ]; then
+		>&2 echo "Server dir $server_dir already exists"
+		exit 1
+	fi
 	# Test extracting $minecraft_zip partially quietly
 	unzip -tq "$minecraft_zip"
 	trap 'sudo rm -rf "$server_dir"' ERR
 	sudo unzip "$minecraft_zip" -d "$server_dir"
 	echo "$current_ver" > "$server_dir/version"
 	sudo chown -R mc:nogroup "$server_dir"
+	echo "@@@ Don't forget to edit $server_dir/server.properties @@@"
 fi
